+++ sls-20201230/alibabacloud_sls20201230/models/_get_logs_v2request.py
@@ -20,43 +20,48 @@ class GetLogsV2Request(DaraModel):
         to: int = None,
         topic: str = None,
     ):
-        # For a scan or phrase query, specifies whether to page forward or backward.
+        # Specifies whether to page forward or backward for scan or phrase queries.
         self.forward = forward
-        # The start of the time range to query. The value is the log time that was specified when the log was written.
+        # The start time of the query. This time refers to the log time specified when log data is written.
         # 
-        # The time range is a left-closed right-open interval. This means the range includes the start time but not the end time. If the from and to values are the same, the interval is invalid and an error is returned. The value is a UNIX timestamp that represents the number of seconds since 00:00:00 UTC on January 1, 1970.
+        # The time range defined by the from and to request parameters follows the left-closed, right-open principle. The time range includes the start time but excludes the end time. If the values of from and to are the same, the time range is invalid and the function returns an error.
+        # The value is a UNIX timestamp representing the number of seconds that have elapsed since January 1, 1970, 00:00:00 UTC.
         # 
         # This parameter is required.
         self.from_ = from_
-        # Specifies whether to highlight the results.
+        # Specifies whether to enable highlighting.
         self.highlight = highlight
-        # Specifies whether to enable nanosecond-level sorting.
+        # Specifies whether to enable nanosecond-precision ordering.
         self.is_accurate = is_accurate
-        # The maximum number of logs to return. This parameter is valid only if the query parameter contains a query statement. The value must be an integer from 0 to 100. The default value is 100.
+        # The maximum number of logs to return in the request. This parameter is valid only when the query parameter is a query statement (not an analytic statement). Minimum value: 0. Maximum value: 100. Default value: 100.
         self.line = line
-        # The line number from which to start the query. This parameter is valid only if the query parameter contains a query statement. The default value is 0.
+        # The start row of the query. This parameter is valid only when the query parameter is a query statement (not an analytic statement). The value starts from 0. Default value: 0.
         self.offset = offset
-        # Specifies whether to enable enhanced SQL. The default value is false.
+        # Specifies whether to enable Dedicated SQL. Disabled by default.
         self.power_sql = power_sql
         # The query statement or analytic statement. For more information, see [Query overview](https://help.aliyun.com/document_detail/43772.html) and [Analysis overview](https://help.aliyun.com/document_detail/53608.html).
         # 
-        # To use the Exclusive SQL feature, add set session parallel_sql=true; to the analytic statement in the query parameter. Example: \\* | set session parallel_sql=true; select count(\\*) as pv.
+        # Add set session parallel_sql=true; to the analytic statement in the query parameter to use Dedicated SQL. Example: * | set session parallel_sql=true; select count(*) as pv.
         # 
-        # Note: If the query parameter contains an analytic statement (SQL statement), the line and offset parameters are invalid. Set them to 0. Use the LIMIT clause in the SQL statement for paging. For more information, see Paginate query and analysis results.
+        # Note: When the query parameter contains an analytic statement (SQL statement), the line and offset parameters of this API are invalid. Set them to 0. Use the LIMIT syntax in the SQL statement for pagination. For more information, see Display query and analysis results by page.
         self.query = query
-        # Specifies whether to return logs in descending order of their timestamps. The precision is at the minute level.
+        # Specifies whether to return logs in descending order of log timestamps, accurate to the minute level. This parameter is valid only when the query parameter is a query statement (not an analytic statement).
         # 
-        # true: Returns logs in descending order of their timestamps. false (default): Returns logs in ascending order of their timestamps. Note: If the query parameter contains a query statement, the reverse parameter is valid and specifies the sorting order. If the query parameter contains a query and analysis statement, the reverse parameter is invalid. The sorting order is specified by the ORDER BY clause in the analytic statement. If ORDER BY is asc (default), logs are sorted in ascending order. If ORDER BY is desc, logs are sorted in descending order.
+        # - true: Returns logs in descending order of log timestamps.
+        # - false (default): Returns logs in ascending order of log timestamps.
+        # 
+        # To sort results in an analytic statement, use the ORDER BY syntax. If ORDER BY is set to asc (default), the results are sorted in ascending order. If ORDER BY is set to desc, the results are sorted in descending order.
         self.reverse = reverse
         # The query parameter.
         self.session = session
-        # The end of the time range to query. The value is the log time that was specified when the log was written.
+        # The end time of the query. This time refers to the log time specified when log data is written.
         # 
-        # The time range is a left-closed right-open interval. This means the range includes the start time but not the end time. If the from and to values are the same, the interval is invalid and an error is returned. The value is a UNIX timestamp that represents the number of seconds since 00:00:00 UTC on January 1, 1970.
+        # The time range defined by the from and to request parameters follows the left-closed, right-open principle. The time range includes the start time but excludes the end time. If the values of from and to are the same, the time range is invalid and the function returns an error.
+        # The value is a UNIX timestamp representing the number of seconds that have elapsed since January 1, 1970, 00:00:00 UTC.
         # 
         # This parameter is required.
         self.to = to
-        # The log topic. The default value is double quotation marks ("").
+        # The topic. Default value: empty string.
         self.topic = topic
 
     def validate(self):