+++ sls-20201230/alibabacloud_sls20201230/models/_get_materialized_view_response_body.py
@@ -2,6 +2,7 @@
 # This file is auto-generated, don't edit it. Thanks.
 from __future__ import annotations
 
+from alibabacloud_sls20201230 import models as main_models
 from darabonba.model import DaraModel
 
 class GetMaterializedViewResponseBody(DaraModel):
@@ -13,25 +14,21 @@ class GetMaterializedViewResponseBody(DaraModel):
         name: str = None,
         original_sql: str = None,
         start_time: int = None,
+        status: main_models.GetMaterializedViewResponseBodyStatus = None,
         ttl: int = None,
     ):
-        # The time interval for data aggregation, in minutes.
         self.agg_interval_mins = agg_interval_mins
-        # Specifies whether the materialized view is active. Valid values are `true` and `false`.
         self.enabled = enabled
-        # The destination logstore that stores the aggregated data.
         self.logstore = logstore
-        # The name of the materialized view.
         self.name = name
-        # The SQL query that defines the materialized view.
         self.original_sql = original_sql
-        # The time when the materialized view was created. The value is a Unix timestamp.
         self.start_time = start_time
-        # The time to live (ttl) of the data in the destination logstore, in days.
+        self.status = status
         self.ttl = ttl
 
     def validate(self):
-        pass
+        if self.status:
+            self.status.validate()
 
     def to_map(self):
         result = dict()
@@ -55,6 +52,9 @@ class GetMaterializedViewResponseBody(DaraModel):
 
         if self.start_time is not None:
             result['startTime'] = self.start_time
+
+        if self.status is not None:
+            result['status'] = self.status.to_map()
 
         if self.ttl is not None:
             result['ttl'] = self.ttl
@@ -81,8 +81,55 @@ class GetMaterializedViewResponseBody(DaraModel):
         if m.get('startTime') is not None:
             self.start_time = m.get('startTime')
 
+        if m.get('status') is not None:
+            temp_model = main_models.GetMaterializedViewResponseBodyStatus()
+            self.status = temp_model.from_map(m.get('status'))
+
         if m.get('ttl') is not None:
             self.ttl = m.get('ttl')
+
+        return self
+
+class GetMaterializedViewResponseBodyStatus(DaraModel):
+    def __init__(
+        self,
+        last_run_error: str = None,
+        last_run_time: int = None,
+        max_cursor_time: int = None,
+    ):
+        self.last_run_error = last_run_error
+        self.last_run_time = last_run_time
+        self.max_cursor_time = max_cursor_time
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.last_run_error is not None:
+            result['lastRunError'] = self.last_run_error
+
+        if self.last_run_time is not None:
+            result['lastRunTime'] = self.last_run_time
+
+        if self.max_cursor_time is not None:
+            result['maxCursorTime'] = self.max_cursor_time
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('lastRunError') is not None:
+            self.last_run_error = m.get('lastRunError')
+
+        if m.get('lastRunTime') is not None:
+            self.last_run_time = m.get('lastRunTime')
+
+        if m.get('maxCursorTime') is not None:
+            self.max_cursor_time = m.get('maxCursorTime')
 
         return self